update trifunction test refactor point class
